fix: don’t add playlist query param on topics
There’s a YouTube bug where playlists play the wrong episode if the index is over 200. This means older LWJ episodes will start playing the wrong episode if we add the playlist to the embed code.

There might be a better fix, but for now this drops the playlist off topic listings, which have a high likelihood of breaking older episodes.

fixes #49